+++ src/components/experience/experience.css
@@ -27,7 +27,7 @@
 .experience__content {
     display: grid;
     grid-template-columns: 1fr 1fr;
-    row-gap: 2rem;
+    row-gap: 1rem;
 }
 
 .experience__details {
