+++ src/components/header/header.css
@@ -1,6 +1,5 @@
 header {
     height: 100vh;
-    padding-top: 3rem;
     overflow: hidden;
 }
 
@@ -45,11 +44,11 @@ header {
 /*====================== ME =====================*/
 .me {
     background: linear-gradient(var(--color-primary), transparent);
-    width: 18.5rem;
+    width: 23.5rem;
     margin: auto;
     left: calc(50% - 11rem);
-    margin-top: 4rem;
-    border-radius: 12rem 12rem 0 0 ;
+    margin-top: 2rem;
+    border-radius: 50%;
     overflow: hidden;
     /* padding: 5rem 1.5rem 1.5rem 1.5rem; */
 }
