+++ src/components/portfolio/Portfolio.js
@@ -18,7 +18,7 @@ const Portfolio = () => {
                     <div className='portfolio__item-image'>
                         <img src={IMG1} alt='Portfolio Image1'/>
                     </div>
-                    <h3>Juego hecho en Angular</h3>
+                    <h3>Tic-tac-toe Game</h3>
                     <div className='portfolio__item-cta'>
                         <a href='https://github.com/MateoDuque03/triqui-game' className='btn'>GitHub</a>
                         <a href='https://triqui-game-git-master.mateoduque03.vercel.app/' className='btn btn-primary' target='_blank' rel="noopener noreferrer">Live Demo</a>
@@ -28,20 +28,20 @@ const Portfolio = () => {
                     <div className='portfolio__item-image'>
                         <img src={IMG2} alt='Portfolio Image1'/>
                     </div>
-                    <h3>This is a portfolio item title</h3>
+                    <h3>React Crypto Price</h3>
                     <div className='portfolio__item-cta'>
-                        <a href='https://github.com' className='btn'>GitHub</a>
-                        <a href='https://github.com' className='btn btn-primary' target='_blank' rel="noopener noreferrer">Live Demo</a>
+                        <a href='https://github.com/MateoDuque03/react-price-crypto' className='btn'>GitHub</a>
+                        <a href='https://react-price-crypto.vercel.app/' className='btn btn-primary' target='_blank' rel="noopener noreferrer">Live Demo</a>
                     </div>
                 </article>
                 <article className='portfolio__item'>
                     <div className='portfolio__item-image'>
                         <img src={IMG3} alt='Portfolio Image1'/>
                     </div>
-                    <h3>This is a portfolio item title</h3>
+                    <h3>React Admin Task</h3>
                     <div className='portfolio__item-cta'>
-                        <a href='https://github.com' className='btn'>GitHub</a>
-                        <a href='https://github.com' className='btn btn-primary' target='_blank' rel="noopener noreferrer">Live Demo</a>
+                        <a href='https://github.com/MateoDuque03/react-admin-task' className='btn'>GitHub</a>
+                        <a href='https://react-admin-task.vercel.app/' className='btn btn-primary' target='_blank' rel="noopener noreferrer">Live Demo</a>
                     </div>
                 </article>
                 {/* <article className='portfolio__item'>
